requirements.txt
requests
bitfinex-tencars
pandas
aiohttp[speedups]
ujson
